move high-level interface over to its own file
